+++ examples/resources/sdwan_mirror_list_policy_object/import.sh
@@ -1,1 +1,0 @@
-terraform import sdwan_mirror_list_policy_object.example "f6b2c44c-693c-4763-b010-895aa3d236bd"