+++ docs/guides/changelog.md
@@ -14,8 +14,9 @@ description: |-
 - Add `sdwan_attach_feature_device_template` resource
 - Add `sdwan_vpn_list_policy_object` resource and data source
 - Add `sdwan_prefix_list_policy_object` resource and data source
-- Add `sdwan_policer_list_policy_object` resource and data source
-- Add `sdwan_mirror_list_policy_object` resource and data source
+- Add `sdwan_policer_policy_object` resource and data source
+- Add `sdwan_mirror_policy_object` resource and data source
+- Add `sdwan_class_map_policy_object` resource and data source
 
 ## 0.1.0